feat(nika-runtime,nika-cli): the production composer + launch gates descend
The run-cluster DESCENT wave, third leg (compute descends, render
stays). The composition root is COMPUTE with zero render coupling —
its seam measured clean: six items cross the boundary
(`ProdRuntime` · `production_runtime` · `RuntimeCapabilities` ·
`capabilities_of` · the fs/net boundary pair · `config_from_env`),
and every edge it drags is strictly downward and acyclic (verified:
no effect crate depends back on nika-runtime).

To `nika-runtime`:
- `compose` — the production composition root, whole: `ProdRuntime` ·
  `production_runtime` (the real fs · http · clock · subprocess ·
  provider registry with env-resolved keys · the OS sandbox pair) ·
  `RegistryProvider` · `StderrEmitter` · the env ladders · the
  `EnvFileSecretResolver` · the `capabilities_of` boundary derivation.
  Every embedder (cli today · daemon/serve/sdk tomorrow) now composes
  the same runtime instead of hand-wiring thirteen crates. The runtime
  CORE stays seam-generic — the new edges serve the composer module
  only, the crate's own code keeps ZERO tokio edge, and the executor
  stays the embedder's (the crate spec §2 amendment records this).
- `SystemStamper` joins the stamper family in `stamp` (UUIDv7 + wall
  time · the live lane beside DeterministicStamper).
- `admit` grows the launch-gate family: `scope_to_task` (the `--task`
  ancestor-cone cut — DAG assembly) + `floor_refusal` +
  `unbounded_breakdown` (the `--max-cost-usd` floor — refuse BEFORE
  any spend, the required_inputs_refusal posture).

Deliberately STAYS in nika-cli: `child_runner.rs` — it speaks the
journal's concrete `TraceFileSink` + `TRACE_DIR`, and the journal's
home is L4 (nika-dap); an L3 crate cannot reach up for it. A
journal-injected child-runner seam is real design work, owed its own
ADR, not this wave.

Consumers point at the new home directly (no re-export shims): run's
mod.rs + child_runner + probe + explain_file rewire to
`nika_runtime::compose`; the cli sheds the descended deps (verb
crates + sha2 → dev-only for the e2e rehearsals · fs/clock/exec-runner/
sandbox pair pruned · uuid → dev-only fixtures).

Zero behavior change: the code moves verbatim (docs and tests ride —
24 tests relocate, cli 292→268, runtime 190→214 · the crate spec §2
amendment + the deny.toml wrapper note record the doctrine).
