+++ crates/nika-runtime/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name          = "nika-runtime"
-description   = "The L3 orchestrator — wave-ordered DAG execution of a CHECKED workflow through the four verb crates, the one event-emission site per verb path (INV-024). The first L3 crate. Spec: docs/crate-specs/nika-runtime.md."
+description   = "The L3 orchestrator — wave-ordered DAG execution of a CHECKED workflow through the four verb crates, the one event-emission site per verb path (INV-024). Also the production composer (compose.rs · the real-effect wiring every embedder shares · descended from nika-cli's run verb 2026-07-22 — compute descends, render stays) and the launch-gate family (admit.rs: required inputs · --task cone · budget floor). The first L3 crate. Spec: docs/crate-specs/nika-runtime.md."
 version.workspace      = true
 edition.workspace      = true
 license.workspace      = true
@@ -24,6 +24,19 @@ nika-verb-infer  = { path = "../nika-verb-infer", version = "0.105.0" }
 nika-verb-exec   = { path = "../nika-verb-exec", version = "0.105.0" }
 nika-verb-invoke = { path = "../nika-verb-invoke", version = "0.105.0" }
 nika-verb-agent  = { path = "../nika-verb-agent", version = "0.105.0" }
+# The production composer (descended from nika-cli's run verb 2026-07-22 ·
+# compute descends, render stays): compose.rs wires the REAL effects into
+# the generic Runtime for every embedder. The runtime CORE stays
+# seam-generic — these edges serve the composer module only; the crate's
+# own code keeps ZERO tokio edge (the effect crates wrap their own).
+nika-builtin     = { path = "../nika-builtin", version = "0.105.0" }      # BuiltinDispatcher — the tool plane (moved up from dev-deps)
+nika-providers   = { path = "../nika-providers", version = "0.105.0" }    # ProviderRegistry — resolve provider/model + keys (moved up from dev-deps)
+nika-fs          = { path = "../nika-fs", version = "0.105.0" }           # TokioFs — the builtin tool plane's fs effect
+nika-clock       = { path = "../nika-clock", version = "0.105.0" }        # SystemClock — backoff + timeout + the date builtin
+nika-http        = { path = "../nika-http", version = "0.105.0" }         # ReqwestHttp — the fetch + provider clients
+nika-exec-runner = { path = "../nika-exec-runner", version = "0.105.0" }  # TokioShell — the exec verb's subprocess runner
+nika-sandbox-seatbelt = { path = "../nika-sandbox-seatbelt", version = "0.105.0" }  # the macOS exec jail (ADR-095 L6)
+nika-sandbox-landlock = { path = "../nika-sandbox-landlock", version = "0.105.0" }  # the Linux exec jail (ADR-095 L6 · bwrap)
 futures-util     = { workspace = true }   # buffered(k) intra-wave concurrency + select (no executor · async rides the injected seams)
 jaq-core         = { workspace = true }   # output: named bindings — THE data language (the same jaq nika:jq runs · spec 04 §binding)
 jaq-std          = { workspace = true }   # jq stdlib filters (map · select · add …) over a binding's raw output
@@ -38,8 +51,6 @@ uuid             = { workspace = true }
 
 [dev-dependencies]
 nika-kernel-mock = { path = "../nika-kernel-mock", version = "0.105.0" }  # MockClock · MockShell · MockToolExecutor · MockToolDefinitionProvider
-nika-providers   = { path = "../nika-providers", version = "0.105.0" }    # ProviderRegistry mock/echo (the tests' InferVerb seam · the composer owns it in prod)
-nika-builtin     = { path = "../nika-builtin", version = "0.105.0" }      # the REAL BuiltinDispatcher over mock fs/http — the adversarial suite's fs-boundary plane (L1.5 ← L3, downward)
 nika-pack        = { path = "../nika-pack", version = "0.105.0" }          # error_codes() — the spec-plane codes the runtime emits pin to the embedded canon
 bytes            = { workspace = true }   # canned HttpResponse bodies (the F1 capturing http seam)
 insta            = { workspace = true }
