+++ deny.toml
@@ -74,7 +74,7 @@ wrappers = [
   "nika-verb-agent",       # L2 verb (s12): tokio dev-dep for #[tokio::test] loop tests over the three mock seams (zero runtime tokio dep · inference + dispatch ride the injected providers)
   "nika-builtin",          # L1.5 (s16): tokio `rt` for spawn_blocking — nika:jq's sync jaq eval runs off the async executor (model-controlled cost · the nika-ocr/a11y pattern) + #[tokio::test] over the mock seams
   "nika-extract",          # L1.5 (s17): tokio `rt` for spawn_blocking — the HTML/feed parse of a model-fetched body (up to 64 MiB) runs off the async executor (same pattern) + #[tokio::test]
-  "nika-runtime",          # L3 orchestrator (s18): tokio dev-dep for the #[tokio::test] conformance floor over the four verb crates (zero runtime tokio dep · async rides the injected verb seams)
+  "nika-runtime",          # L3 orchestrator (s18): tokio dev-dep for the #[tokio::test] conformance floor over the four verb crates — the crate's own code keeps ZERO tokio edge (async rides the injected verb seams); the descended production composer (2026-07-22) reaches tokio only TRANSITIVELY through the wrapper effect crates below (nika-fs/http/exec-runner), never directly
   "nika-dap",              # L4 forensics (the trust plane · 2026-07-20): tokio dev-dep ONLY — the anchor submit-path tests park a current-thread executor around the HttpPostDyn seam (zero runtime tokio dep · the L4 shim owns the production executor)
   # Transitive consumers via reqwest → hyper stack (allowed through nika-catalog-verify)
   # tower-http joined via chromiumoxide's reqwest 0.13 tree (nika-browser M2.5).
